# Improve performance of external (UDR) functions