Test preparation:
```
recreate function div_udr (
    n1 integer,
    n2 integer
) returns double precision
    external name 'udf_compat!UC_div'
    engine udr;

create domain d_int_check integer check (value > 0);
create domain d_double_check double precision check (value >= 0);

recreate function div_udr2 (
    n1 d_int_check,
    n2 d_int_check
) returns d_double_check
    external name 'udf_compat!UC_div'
    engine udr;
```

Comparison:
```
set stats on;

-- 4.4s -> 2.3s
execute block
as
    declare n integer = 5000000;
    declare x integer;
begin
    while (n > 0) do
    begin
        n = n - 1;
        x = div_udr(3, 2);
    end
end;

-- 5.1s -> 2.6s
execute block
as
    declare n integer = 5000000;
    declare x integer;
begin
    while (n > 0) do
    begin
        n = n - 1;
        x = div_udr2(3, 2);
    end
end;
```