@sim1984 2024-01-31
Does this only improve the performance of external functions only, or is there an effect for procedures/triggers too?