@asfernandes 2024-01-31
> Does this only improve the performance of external functions only, or is there an effect for procedures/triggers too?

It's mostly about functions only.

There could be a marginal improvement with procedures with parameters with constraints, but I didn't tested it, as procedure/trigger implementation could also be greater improved, but is much more difficult.